Add font outline support (#83)
* Add font outline support

* Use GL_NEAREST instead

* Nvm

* Add Conan profile init

* Update main.yml

* Fix glm

* Update main.yml

* Update main.yml

* Replace long long with int64_t

* Update vector3.h

* Update vector2.h

* Update vector4.h

* Update color.h

---------

+++ game/include/mainframe/game/engine.h
@@ -25,7 +25,7 @@ namespace mainframe::game {
 		virtual void pollEvents();
 
 		// called on a fixed timestep
-		virtual void update(float deltaTime, long long gameTime);
+		virtual void update(float deltaTime, int64_t gameTime);
 
 		// called acordingly with the FPS lock
 		virtual void draw(const double alpha);
