+++ ui/tests/simplemenu/src/main.cpp
@@ -87,7 +87,7 @@ public:
 		window.swapBuffer();
 	}
 
-	virtual void update(float deltaTime, long long gameTime) override {
+	virtual void update(float deltaTime, int64_t gameTime) override {
 		Window::pollEvents();
 		scene->update(deltaTime);
 	}
